docs: add SectionOfItem story, renaming incorrect filename
