+++ packages/gatsby-theme-project-portal/src/components/SectionOfItem.stories.tsx
@@ -1,0 +1,25 @@
+import type { Meta, StoryObj } from "@storybook/react"
+import { SectionOfItem } from "./SectionOfItem"
+
+const meta: Meta<typeof SectionOfItem> = {
+  component: SectionOfItem,
+  tags: ["autodocs"],
+}
+
+export default meta
+
+type Story = StoryObj<typeof SectionOfItem>
+
+const text = `Lorem ipsum *dolor* sit amet, consectetur adipiscing elit, sed do eiusmod tempor 
+incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
+exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure 
+dolor in reprehenderit in voluptate __velit esse__ cillum dolore eu fugiat nulla pariatur. 
+Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit 
+anim id est laborum.`
+
+export const Primary: Story = {
+  args: {
+    label: "Heading",
+    value: text,
+  },
+}
