+++ packages/gatsby-theme-project-portal/src/components/SectionOfItem.tsx
