+++ packages/gatsby-theme-project-portal/src/components/index.tsx
@@ -6,7 +6,7 @@ export { ProjectStatus } from "./ProjectStatus"
 export { MarkdownText } from "./MarkdownText"
 export { Feature } from "./Feature"
 export { ShareProject } from "./ShareProject"
-export { SectionOfItem } from "./SectionofItem"
+export { SectionOfItem } from "./SectionOfItem"
 export { Card } from "./Card"
 export type { CardProps } from "./Card"
 export { Cards } from "./Cards"
